+++ src/player/musicPlayer/ResolverClient.js
@@ -11,12 +11,24 @@ export class ResolverClient {
     return this.host._resolveSpotifyTrack(url, requestedBy);
   }
 
+  resolveAppleTrack(url, requestedBy) {
+    return this.host._resolveAppleTrack(url, requestedBy);
+  }
+
   resolveSpotifyCollection(url, requestedBy) {
     return this.host._resolveSpotifyCollection(url, requestedBy);
   }
 
+  resolveAppleCollection(url, requestedBy) {
+    return this.host._resolveAppleCollection(url, requestedBy);
+  }
+
   resolveSpotifyByGuess(url, requestedBy) {
     return this.host._resolveSpotifyByGuess(url, requestedBy);
+  }
+
+  resolveAppleByGuess(url, requestedBy) {
+    return this.host._resolveAppleByGuess(url, requestedBy);
   }
 
   resolveFallback(url, requestedBy, source) {