+++ src/player/musicPlayer/sourceMethods.js
@@ -1,3 +1,4 @@
+import { appleMethods } from './appleMethods.js';
 import { audiusMethods } from './audiusMethods.js';
 import { deezerMethods } from './deezerMethods.js';
 import { soundcloudMethods } from './soundcloudMethods.js';
@@ -6,6 +7,7 @@ import { trackFactoryMethods } from './trackFactoryMethods.js';
 import { urlResolverMethods } from './urlResolverMethods.js';
 
 export const sourceMethods = {
+  ...appleMethods,
   ...trackFactoryMethods,
   ...audiusMethods,
   ...soundcloudMethods,